% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bgg}
\alias{bgg}
\title{BoardGameGeek Game Data}
\format{
An object of class \verb{<geek_item>} (a named list of data frames)
}
\usage{
bgg
}
\description{
A dataset containing 611 games acquired
from the \href{https://boardgamegeek.com}{BoardGameGeek} (BGG) database.
}
\section{info}{


This dataset includes basic game info.
\itemize{
\item \strong{id} int, a unique game id.
\item \strong{name} chr, primary name.
\item \strong{yearpublished} int
\item \strong{description} chr
\item \strong{minage} int
\item \strong{minplayers} int
\item \strong{maxplayers} int
\item \strong{playingtime} int
\item \strong{minplaytime} int
\item \strong{maxplaytime} int
\item \strong{image} chr, cover image URL.
\item \strong{seriescode} logi
\item \strong{releasedate} logi
}
}

\section{altnames}{


This dataset includes alternative game names.
\itemize{
\item \strong{id} int, a unique game id.
\item \strong{altname} chr, game name.
\item \strong{type} chr, indicates whether the name is a primary or alternate name.
}
}

\section{numplayers}{


This dataset includes poll results on the best number of players for a game.
\itemize{
\item \strong{id} int, a unique game id.
\item \strong{numplayers} chr, number of players, can be \verb{4+} and similar string values.
\item \strong{Best} int, how many times this number of players was voted "Best".
\item \strong{Recommended} int, how many times this number of players was
voted "Recommended".
\item \strong{Not Recommended} int, how many times this number of players was voted
"Not Recommended".
}
}

\section{playerage}{


This dataset includes poll results on the most appropriate player age.
\itemize{
\item \strong{id} int, a unique game id.
\item \strong{playerage} chr, can be mostly numerical but can be \code{"21 and up"}.
\item \strong{votes} int how many times this age was upvoted.
}
}

\section{language}{


This dataset includes poll results on the degree of language requirements.
\itemize{
\item \strong{id} int, a unique game id.
\item \strong{langlevel} chr, one of four values ordered from no in-game text to
unplayable in another language.
\item \strong{votes} int, how many times this level upvoted.
}
}

\section{links}{


This dataset includes information on related items, such as mechanics.
\itemize{
\item \strong{id} int, a unique game id
\item \strong{link_id} int, related item id.
\item \strong{link_type} chr, related item type, such \code{"mechanic"} or \code{"category"}.
\item \strong{link_name} chr, related item name.
}
}

\section{stats}{


This dataset includes game statistics.
\itemize{
\item \strong{id} int, a unique game id.
\item \strong{usersrated} int, voter count.
\item \strong{average} num, average rating.
\item \strong{bayesaverage} num, a combination of the average rating and thirty
additional average ratings across the whole site. This is intended
to prevent a new or rare game with only a few high ratings from taking
the top spots. See
\href{https://boardgamegeek.com/thread/71129/what-bayesian-average}{BGG forum.}
\item \strong{stddev} num
\item \strong{median} int
\item \strong{owned} int, number of people who sold a game.
\item \strong{trading} int, number of people who seek to sell a game.
\item \strong{wanting} int, number of people who seek to buy a game.
\item \strong{wishing} int
\item \strong{numcomments} int
\item \strong{numweights} int
\item \strong{averageweight} num, a measure of game complexity.
}
}

\keyword{datasets}
