% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{geek_item}
\alias{geek_item}
\alias{game_info}
\alias{game_altnames}
\alias{game_numplayers}
\alias{game_playerage}
\alias{game_language}
\alias{game_stats}
\alias{game_links}
\alias{as_xml_nodeset}
\alias{is_geek_item}
\title{Geek Game Data}
\usage{
game_info(xml)

game_altnames(xml)

game_numplayers(xml)

game_playerage(xml)

game_language(xml)

game_stats(xml)

game_links(xml)

as_xml_nodeset(x)

is_geek_item(x)
}
\arguments{
\item{xml}{A xml nodeset of class \code{geek_item}.}

\item{x}{An object to be tested or converted to an xml nodeset.}
}
\value{
A tibble (data frame)
}
\description{
A collection of functions to deal with geek items, produced by \code{\link[=get_game]{get_game()}}.
}
\details{
Most functions in this family parse game data and return tibbles (data
frames). \code{is_geek_item()} returns TRUE if an object is produced by \code{\link[=get_game]{get_game()}}
and contains game data. \code{as_xml_nodeset} strips the \code{"geek_item"} class.
}
