% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeks.R
\docType{package}
\name{geeks}
\alias{geeks}
\title{geeks: Geek API client}
\description{
The purpose of \code{geeks} is to fetch data from three famous geek websites
using \href{https://boardgamegeek.com/wiki/page/BGG_XML_API2}{API2}.
}
\details{
Geek websites:
\itemize{
\item https://www.boardgamegeek.com
\item https://rpggeek.com
\item https://videogamegeek.com
}
}
