% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-families.R
\name{get_families}
\alias{get_families}
\title{Get Item Family Data}
\usage{
get_families(id)
}
\arguments{
\item{id}{A vector of family id numbers.}
}
\value{
A tibble. To extract the raw HTTP response, call \code{attr(x, "res")}.
}
\description{
Request item family data.
}
