% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-game.R
\name{get_game}
\alias{get_game}
\title{Get Game Data}
\usage{
get_game(
  id,
  domain = c("boardgame", "rpg", "videogame"),
  data = c("stats", "versions", "historical", "marketplace", "comments",
    "ratingcomments", "videos"),
  silent = TRUE
)
}
\arguments{
\item{id}{A vector of game id numbers.}

\item{domain}{One of \code{"boardgame"}, \code{"rpg"} or \code{"videogame"}. Determines
which domain the request will go to. Defaults to \code{"boardgame"}.}

\item{data}{What kinds of data to retrieve (besides basic info). See
\href{https://boardgamegeek.com/wiki/page/BGG_XML_API2#toc3}{API2 documentation}
for details.}

\item{silent}{Whether to print the URL on request.}
}
\value{
An xml2 nodeset of class "geek_item"
}
\description{
Request game data from one of the geek sites.
}
