% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-games.R
\name{get_games}
\alias{get_games}
\title{Get Game Data}
\usage{
get_games(
  id,
  add = c("stats", "versions", "marketplace", "videos", "comments", "ratingcomments"),
  page_start = 1,
  page_n = 1,
  delay_s = 0.5
)
}
\arguments{
\item{id}{A vector of game id numbers.}

\item{add}{One or multiple of \code{"stats"}, \code{"versions"}, \code{"marketplace"},
\code{"videos"}, \code{"comments"}, \code{"ratingcomments"}.

\strong{Note.} The API will not return both \code{comments} and \code{ratingcomments}
in the same call. If you request both, you will only get \code{comments}.}

\item{page_start}{Page number from which to start paging. Defaults to \code{1}.}

\item{page_n}{How many pages to fetch.}

\item{delay_s}{How many seconds to wait between requests. Defaults to \code{0.5}.}
}
\value{
A tibble. To extract the raw HTTP response, call \code{attr(x, "res")}.
}
\description{
Request game data.
}
\examples{

# You don't really need to name the vector, but it makes things clearer.
ids = c(gloomhaven = 174430, agricola = 31260, junkart = 193042)
res = get_games(ids, add = "stats")
res
}
