% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-plays.R
\name{get_plays}
\alias{get_plays}
\title{Get Plays Data}
\usage{
get_plays(id, page_start = 1, page_n = 1, delay_s = 0.5)
}
\arguments{
\item{id}{A game id number.}

\item{page_start}{Starting page.}

\item{page_n}{How many pages to fetch.}

\item{delay_s}{How many seconds to wait between requests. Defaults to \code{0.5}.}
}
\value{
A tibble. To extract the raw HTTP response, call \code{attr(x, "res")}.
}
\description{
Request play data.
}
