% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{susd}
\alias{susd}
\title{Shut Up And Sit Down Recommends}
\format{
A data frame with 2 variables
\itemize{
\item \strong{id} corresponds to the game entry in BoardGameGeek database.
\item \strong{susd_recommends} indicates whether a game is recommended (1) by SUSD
or not (0).
}
}
\source{
https://www.shutupandsitdown.com
}
\usage{
susd
}
\description{
A dataset containing 611 games reviewed
by \href{https://www.shutupandsitdown.com/}{Shut Up And Sit Down} (SUSD).
}
\keyword{datasets}
