% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bgg}
\alias{bgg}
\title{BoardGameGeek Game Data}
\format{
An object of class \verb{<geek_item>} (a named list of data frames)
}
\usage{
bgg
}
\description{
A dataset containing 611 games acquired
from the \href{https://boardgamegeek.com}{BoardGameGeek} (BGG) database.
}
\section{info}{


This dataset includes basic game info.
\itemize{
\item \strong{id} int, a unique game id.
\item \strong{name} chr, primary name.
\item \strong{yearpublished} int
\item \strong{description} chr
\item \strong{minage} int
\item \strong{minplayers} int
\item \strong{maxplayers} int
\item \strong{playingtime} int
\item \strong{minplaytime} int
\item \strong{maxplaytime} int
\item \strong{image} chr, cover image URL.
\item \strong{seriescode} logi
\item \strong{releasedate} logi
}
}

\section{altnames}{


This dataset includes alternative game names.
\itemize{
\item \strong{id} int, a unique game id.
\item \strong{altname} chr, game name.
\item \strong{type} chr, indicates whether the name is a primary or alternate name.
}
}

\section{numplayers}{


This dataset includes poll results on the best number of players for a game.
\itemize{
\item \strong{id} int, a unique game id.
\item \strong{numplayers} chr, number of players, can be \verb{4+} and similar string values.
\item \strong{option} int, whether this number of players was voted \code{"Best"},
\code{"Recommended"} or \code{"Not Recommended"}.
\item \strong{votes} int, how many times was this number of players voted for.
}
}

\section{playerage}{


This dataset includes poll results on the most appropriate player age.
\itemize{
\item \strong{id} int, a unique game id.
\item \strong{playerage} chr, mostly numbers but can be \code{"21 and up"}.
\item \strong{votes} int, how many times this age was upvoted.
}
}

\section{language}{


This dataset includes poll results on the degree of language requirements.
\itemize{
\item \strong{id} int, a unique game id.
\item \strong{langlevel} chr, one of four ordered values ranging from "no in-game text"
to "unplayable in another language".
\item \strong{votes} int, how many times this level upvoted.
}
}

\section{links}{


A list of data frames. Includes information on related items, such
as mechanics, categories, designers or artists. The actual column names
follow the same pattern but depend on what type of link is included.
\itemize{
\item \strong{id} int, a unique game id
\item \strong{link_id} int, related item id.
\item \strong{link_name} chr, related item name.
}
}

\section{stats}{


This dataset includes game statistics.
\itemize{
\item \strong{id} int, a unique game id.
\item \strong{usersrated} int, voter count.
\item \strong{average} num, average rating.
\item \strong{bayesaverage} num, see note below.
\item \strong{stddev} num
\item \strong{median} int
\item \strong{owned} int, how many people currently own this game.
\item \strong{trading} int, how many people want to sell this game.
\item \strong{wanting} int, how many people want to buy this game.
\item \strong{wishing} int, how many people have this game on their wishlist.
\item \strong{numcomments} int
\item \strong{numweights} int
\item \strong{averageweight} num, a measure of game complexity.
}

\strong{bayeaverage.} Games with less than 30 ratings do not have \code{bayesaverage},
so the value is \code{NA}. For games with at least 30 ratings, an average rating
across the whole database is taken, then repeated a number of times (starting
at 500 but can be more) and added to the actual ratings. This is to prevent
games with very few ratings to reach extremely low or high averages. See
\href{https://boardgamegeek.com/thread/71129/what-bayesian-average}{this BGG thread.}
for discussion.
}

\keyword{datasets}
