% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeks.R
\docType{package}
\name{geeks}
\alias{geeks}
\title{geeks: Geek API client}
\description{
The purpose of \code{geeks} is to fetch tidy data from
\href{https://www.boardgamegeek.com}{boardgamegeek},
\href{https://rpggeek.com}{rpggeek} and
\href{https://videogamegeek.com}{videogamegeek}. Under the hood, it is using the
\href{https://github.com/r-lib/httr}{httr} package to query the
\href{https://boardgamegeek.com/wiki/page/BGG_XML_API2}{XMLAPI2}.
}
